Fix renter D3D11 presenter build; don't block StreamHost on renter failure.
Use Vortice D3D11CreateDevice out-params in presenter; allow renter zip step
to fail without skipping StreamHost/FFmpeg publish to droplet.

+++ stream/windows/renter/PCHUB.Renter/Rendering/GpuVideoHost.cs
@@ -102,8 +102,8 @@ public sealed class GpuVideoHost : HwndHost
     {
         if (_presenter is null) return false;
 
-        _streamWidth = raw.Width;
-        _streamHeight = raw.Height;
+        _streamWidth = (int)raw.Width;
+        _streamHeight = (int)raw.Height;
         UpdateLetterboxLayout();
         return _presenter.Present(raw);
     }
